accumulative_hash

Struct AccumulativeHashSet

pub struct AccumulativeHashSet<T>
where
    T: IsAccumulativeHashType + Hash,
{ /* private fields */ }

A set model using HashSet for accumulative hash types.

This is a convenience wrapper around AccumulativeHash to track visited, order-independent paths in a DFS-like traversal. It uses a HashSet with a NoHashHasher to store and compare visited states efficiently, and uses an internal stack to manage the current path for backtracking.

Implementations

impl<T> AccumulativeHashSet<T>
where T: IsAccumulativeHashType + Hash, BuildNoHashHasher<T>: BuildHasher,

pub fn hasher_state(&self) -> &T

Get a reference to the current hasher state.

pub fn visited_count(&self) -> usize

Get a count of visited states.

pub fn contains_path_to<S: Into<T> + Copy>(&self, element: S) -> bool

Checks if an element, if added to the current hash state, would lead to a previously visited state.

This does not mutate the hash state, only checks for presence.

pub fn transverse_to<S: Into<T> + Copy>(&mut self, element: S) -> bool

Advance the hash state by adding an element.

If the resulting state has been seen before, rollback the hash state and return false. Otherwise return true.

This mutates the hash state, but does not mark the state as visited; this is done by [insert].

pub fn backtrack(&mut self) -> bool

Backtrack the last added element, reverting the hash state.

This does NOT mark the state as visited.

pub fn visit_and_backtrack(&mut self) -> bool

Mark the current hash state as visited and backtrack the last added element.

This is a convenience method that combines marking the current state as visited and backtracking.