accumulative_hash

Trait IsAccumulativeHashType

pub trait IsAccumulativeHashType:
    Sized
    + Copy
    + PartialEq
    + Eq
    + BitXor<Output = Self>
    + Shr<Output = Self>
    + WrappingAdd
    + WrappingSub
    + WrappingMul
    + Zero {
    const SEED: Self;
    const SHIFT_CONSTANTS: [Self; 3];
    const MULTIPLIER_CONSTANTS: [Self; 2];
}

Trait defining the properties required for an accumulative hash type.

Only implemented for primitive unsigned integer types: u8, u16, u32, u64, and u128, and the usage of anything below u64 is not recommended for production use due to the high likelihood of collisions.

Required Associated Constants

const SEED: Self

const SHIFT_CONSTANTS: [Self; 3]

const MULTIPLIER_CONSTANTS: [Self; 2]

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.