accumulative_hash

Trait IsAtomicAccumulativeHashType

pub trait IsAtomicAccumulativeHashType: Sync + Send
where
    Self: From<Self::UnderlyingType>,
{
    type UnderlyingType: IsAccumulativeHashType;

    // Required methods
    fn to_underlying(&self, order: Ordering) -> Self::UnderlyingType;
    fn compare_exchange(
        &self,
        current: Self::UnderlyingType,
        new: Self::UnderlyingType,
        success: Ordering,
        failure: Ordering,
    ) -> Result<Self::UnderlyingType, Self::UnderlyingType>;
}

Trait defining the properties required for an atomic accumulative hash type.

This is a sub-trait of IsAccumulativeHashType that adds an associated type representing the underlying non-atomic type.

Required Associated Types

type UnderlyingType: IsAccumulativeHashType

Required Methods

fn to_underlying(&self, order: Ordering) -> Self::UnderlyingType

Convert the atomic type to its underlying type.

fn compare_exchange( &self, current: Self::UnderlyingType, new: Self::UnderlyingType, success: Ordering, failure: Ordering, ) -> Result<Self::UnderlyingType, Self::UnderlyingType>

Stores a value into the atomic integer if the current value is the same as the current value.

The return value is a result indicating whether the new value was written and containing the previous value. On success this value is guaranteed to be equal to current.

compare_exchange takes two Ordering arguments to describe the memory ordering of this operation. success describes the required ordering for the read-modify-write operation that takes place if the comparison with current succeeds. failure describes the required ordering for the load operation that takes place when the comparison fails. Using Ordering::Acquire as success ordering makes the store part of this operation Ordering::Relaxed, and using Ordering::Release makes the successful load Ordering::Relaxed. The failure ordering can only be Ordering::SeqCst, Ordering::Acquire or Ordering::Relaxed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.